UML-kaavion lisäys kuvana ja linkkinä, tapahtumat suoralla yhteydellä osoitteeseen
