--- ticketguru/src/main/java/bugivelhot/ticketguru/repository/TapahtumapaikkaRepository.java
@@ -1,10 +1,0 @@
-package bugivelhot.ticketguru.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import bugivelhot.ticketguru.model.Tapahtumapaikka;
-
-public interface TapahtumapaikkaRepository extends  JpaRepository<Tapahtumapaikka, Long> {
-
-
-}
